+++ Reynj.UnitTests/Linq/UnionTests.cs
@@ -1,0 +1,197 @@
+﻿using System;
+using System.Collections.Generic;
+using FluentAssertions;
+using Reynj.Linq;
+using Xunit;
+
+namespace Reynj.UnitTests.Linq
+{
+    public class UnionTests
+    {
+        [Fact]
+        public void Union_WithNull_IsNotPossible()
+        {
+            // Arrange
+            IEnumerable<Range<int>> ranges = new Range<int>[] { };
+
+            // Act
+            // ReSharper disable once ExpressionIsAlwaysNull
+            Action act = () => ranges.Union(null);
+
+            // Assert
+            act.Should().Throw<ArgumentNullException>()
+                .And.ParamName.Should().Be("second");
+        }
+
+        [Fact]
+        public void Union_AsNull_IsNotPossible()
+        {
+            // Arrange
+            IEnumerable<Range<int>> ranges = null;
+
+            // Act
+            // ReSharper disable once ExpressionIsAlwaysNull
+            Action act = () => ranges.Union(new Range<int>[] { });
+
+            // Assert
+            act.Should().Throw<ArgumentNullException>()
+                .And.ParamName.Should().Be("first");
+        }
+
+        [Theory]
+        [MemberData(nameof(UnionData))]
+        public void Union_ReturnsTheExpectedResult(IEnumerable<Range<int>> first, IEnumerable<Range<int>> second,
+            IEnumerable<Range<int>> expectedUnion)
+        {
+            // Act
+            var unioned = first.Union(second);
+
+            // Assert
+            unioned.Should().BeEquivalentTo(expectedUnion);
+        }
+
+        [Theory]
+        [MemberData(nameof(UnionData))]
+        public void Union_ReturnsTheExpectedResult_OtherWayAround(IEnumerable<Range<int>> first,
+            IEnumerable<Range<int>> second, IEnumerable<Range<int>> expectedUnion)
+        {
+            // Act
+            var unioned = second.Union(first);
+
+            // Assert
+            unioned.Should().BeEquivalentTo(expectedUnion);
+        }
+
+        public static IEnumerable<object[]> UnionData()
+        {
+            // Empty Lists
+            yield return new object[]
+            {
+                new List<Range<int>>(),
+                new List<Range<int>>(),
+                new List<Range<int>>()
+            };
+
+            // A single Range that is the same
+            yield return new object[]
+            {
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 10)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 10)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 10)
+                })
+            };
+
+            // An empty Range
+            yield return new object[]
+            {
+                new List<Range<int>>(new[]
+                {
+                    Range<int>.Empty
+                }),
+                new List<Range<int>>(),
+                new List<Range<int>>()
+            };
+
+            // An empty Range combined with a single Range
+            yield return new object[]
+            {
+                new List<Range<int>>(new[]
+                {
+                    Range<int>.Empty,
+                    new Range<int>(0, 10)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 10)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 10)
+                })
+            };
+
+            // Two touching Ranges
+            yield return new object[]
+            {
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 10),
+                    new Range<int>(10, 20)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 20)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 20)
+                })
+            };
+
+            // Included in the other Range
+            yield return new object[]
+            {
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 20),
+                    new Range<int>(5, 15)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 20)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 20)
+                })
+            };
+
+            // Non-overlapping Ranges
+            yield return new object[]
+            {
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 10)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(20, 30)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 10),
+                    new Range<int>(20, 30)
+                })
+            };
+
+            // Complex
+            yield return new object[]
+            {
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 5),
+                    new Range<int>(3, 10),
+                    new Range<int>(10, 15)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(15, 17),
+                    new Range<int>(18, 25)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 17),
+                    new Range<int>(18, 25)
+                })
+            };
+        }
+    }
+}